requirements.txt
django

gunicorn
django-heroku

djangorestframework
django-rest-swagger
django-rest-auth

django-celery-beat
django-celery-results

django-cors-headers

numpy
pandas